+++ test/serialize_test.dart
@@ -16,8 +16,8 @@ main() {
     app = new Angel(reflector: MirrorsReflector())
       ..get('/foo', ioc(() => {'hello': 'world'}))
       ..get('/bar', (req, res) async {
-        res.contentType = new MediaType('text', 'html');
-        await res.serialize({'hello': 'world'});
+        await res.serialize({'hello': 'world'},
+            contentType: new MediaType('text', 'html'));
       });
     client = new http.Client();